PR review - visual updates - #4712
