+++ ui/app/src/components/RemoteRepositoriesPullDialog/RemoteRepositoriesPullDialog.tsx
@@ -16,79 +16,13 @@
 
 import React, { useState } from 'react';
 import Dialog from '@material-ui/core/Dialog';
-import DialogHeader from '../Dialogs/DialogHeader';
-import DialogBody from '../Dialogs/DialogBody';
-import DialogFooter from '../Dialogs/DialogFooter';
-import { FormattedMessage } from 'react-intl';
-import InputLabel from '@material-ui/core/InputLabel';
-import FormControl from '@material-ui/core/FormControl';
-import Select from '@material-ui/core/Select';
-import MenuItem from '@material-ui/core/MenuItem';
-import { createStyles, makeStyles } from '@material-ui/core/styles';
-import { MergeStrategy } from '../../models/Repository';
-import { pull } from '../../services/repositories';
-import { useActiveSiteId } from '../../utils/hooks';
-import ApiResponse from '../../models/ApiResponse';
-import SecondaryButton from '../SecondaryButton';
-import PrimaryButton from '../PrimaryButton';
-import { isBlank } from '../../utils/string';
+import RemoteRepositoriesPullDialogContainer, {
+  RemoteRepositoriesPullDialogContainerProps
+} from './RemoteRepositoriesPullDialogContainer';
 
-export interface RemoteRepositoriesPullDialogProps {
-  open: boolean;
-  branches: string[];
-  remoteName: string;
-  mergeStrategies: MergeStrategy[];
-  onClose(): void;
-  onPullSuccess?(): void;
-  onPullError?(response: ApiResponse): void;
-}
-
-const useStyles = makeStyles(() =>
-  createStyles({
-    formControl: {
-      marginBottom: '15px'
-    }
-  })
-);
-
-export default function RemoteRepositoriesPullDialog(props: RemoteRepositoriesPullDialogProps) {
-  const { open, branches, remoteName, mergeStrategies, onClose, onPullSuccess, onPullError } = props;
-  const [selectedBranch, setSelectedBranch] = useState('');
-  const [selectedMergeStrategy, setSelectedMergeStrategy] = useState(mergeStrategies[0].key);
-  const classes = useStyles();
-  const siteId = useActiveSiteId();
+export default function RemoteRepositoriesPullDialog(props: RemoteRepositoriesPullDialogContainerProps) {
+  const { open, onClose } = props;
   const [disableQuickDismiss, setDisableQuickDismiss] = useState(false);
-
-  const onChange = (e: any) => {
-    e.persist();
-    if (e.target.name === 'branch') {
-      setSelectedBranch(e.target.value);
-      setDisableQuickDismiss(true);
-    } else if (e.target.name === 'mergeStrategy') {
-      setSelectedMergeStrategy(e.target.value);
-    }
-  };
-
-  const onSubmit = (e) => {
-    e.preventDefault();
-
-    if (!isBlank(selectedBranch)) {
-      pull({
-        siteId,
-        remoteName,
-        remoteBranch: selectedBranch,
-        mergeStrategy: selectedMergeStrategy
-      }).subscribe(
-        () => {
-          onPullSuccess?.();
-        },
-        ({ response }) => {
-          onPullError?.(response.response);
-        }
-      );
-      onClose();
-    }
-  };
 
   return (
     <Dialog
@@ -99,57 +33,7 @@ export default function RemoteRepositoriesPullDialog(props: RemoteRepositoriesPu
       disableBackdropClick={disableQuickDismiss}
       disableEscapeKeyDown={disableQuickDismiss}
     >
-      <form onSubmit={onSubmit}>
-        <DialogHeader title={<FormattedMessage id="words.pull" defaultMessage="Pull" />} onDismiss={onClose} />
-        <DialogBody>
-          <FormControl variant="outlined" fullWidth className={classes.formControl}>
-            <InputLabel id="remoteBranchToPullLabel">
-              <FormattedMessage id="repositories.remoteBranchToPull" defaultMessage="Remote Branch to Pull" />
-            </InputLabel>
-            <Select
-              labelId="remoteBranchToPullLabel"
-              name="branch"
-              value={selectedBranch}
-              onChange={onChange}
-              label={<FormattedMessage id="repositories.remoteBranchToPull" defaultMessage="Remote Branch to Pull" />}
-              fullWidth
-            >
-              {branches.map((branch) => (
-                <MenuItem key={branch} value={branch}>
-                  {branch}
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
-          <FormControl variant="outlined" fullWidth>
-            <InputLabel id="mergeStrategyLabel">
-              <FormattedMessage id="repositories.mergeStrategyLabel" defaultMessage="Merge Strategy" />
-            </InputLabel>
-            <Select
-              labelId="mergeStrategyLabel"
-              name="mergeStrategy"
-              value={selectedMergeStrategy}
-              onChange={onChange}
-              label={<FormattedMessage id="repositories.mergeStrategyLabel" defaultMessage="Merge Strategy" />}
-              fullWidth
-            >
-              {mergeStrategies.map((strategy) => (
-                <MenuItem key={strategy.key} value={strategy.key}>
-                  {strategy.value}
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
-        </DialogBody>
-        <DialogFooter>
-          <SecondaryButton onClick={onClose}>
-            <FormattedMessage id="words.cancel" defaultMessage="Cancel" />
-          </SecondaryButton>
-          <PrimaryButton type="submit" disabled={isBlank(selectedBranch)}>
-            <FormattedMessage id="words.ok" defaultMessage="Ok" />
-          </PrimaryButton>
-        </DialogFooter>
-      </form>
+      <RemoteRepositoriesPullDialogContainer {...props} setDisableQuickDismiss={setDisableQuickDismiss} />
     </Dialog>
   );
 }
